+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Update `mio` dependency
 - [server] Fix some leaks when dropping the display when using the rust implementation
+- [protocols] Update wayland-protocols and wlroots-protocols.
 
 ## 0.21.7 -- 2018-11-28